refactor: move 'trackerModule' into a separate file
