+++ src/plugins/tracker.ts
@@ -1,16 +1,5 @@
 import { App, inject } from "vue";
-import firebase from "firebase";
-
-export interface ITrackConfig {
-  user: firebase.User | null;
-}
-
-export interface ITracker {
-  events: {
-    [key: string]: (route: any, options?: unknown) => void;
-  };
-  track(eventName: string, data?: any): void;
-}
+import { createTrackerModule, ITrackConfig, ITracker } from "./trackerModule";
 
 const TRACKER_SYMBOL = Symbol("tracker");
 
@@ -19,46 +8,7 @@ const TRACKER_SYMBOL = Symbol("tracker");
 
 export default {
   install: (app: App, config: ITrackConfig) => {
-    const trackerModule: ITracker = {
-      events: {
-        /* TODO: limit 'any' type of 'options' */
-        routeChange: (data: any) => {
-          const { route, enteredFrom, exceptions } = data;
-          if (
-            !route ||
-            exceptions.includes(route) ||
-            exceptions.includes(enteredFrom)
-          )
-            return;
-          const eventInfo = {
-            route,
-            enteredFrom,
-            enteredBy: config.user?.uid,
-            timestamp: new Date().toISOString(),
-          };
-          firebase
-            .database()
-            .ref("analytics")
-            .child("routeChanges")
-            .push(eventInfo);
-        },
-
-        customEvent: (eventName, data: any) => {
-          const uid = config.user?.uid;
-          const info = { ...data, uid };
-
-          firebase.database().ref("analytics").child(eventName).push(info);
-        },
-      },
-
-      track(eventName, data) {
-        if (Object.prototype.hasOwnProperty.call(this.events, eventName)) {
-          return this.events[eventName](data);
-        }
-
-        this.events.customEvent(eventName, data);
-      },
-    };
+    const trackerModule = createTrackerModule(config);
 
     app.provide(TRACKER_SYMBOL, trackerModule);
     app.config.globalProperties.$tracker = trackerModule;
